Load friend's cover from backend
Before the friend page was using Facebook's API to request the cover
source in order to use on the app, now we use the URL provided by our
own backend reducing the number of requests and complexity of our code.
Also configures $httpProvider to cache GET request, reducing load times
and the number of request made on every page change.